[dynamicIO] implement validation using same methods as prerender
In Dev we simulate prerenders to provide useful information about violations of dynamicIO rules. To achieve this we run a simulated prerender on every page request. In a prior update we changed the implementation of dynamicIO prerenders and this change updates the implementation to match the actual prerender implementation as much as possible. Some things are intentionally different such as using a hmr hash to re-use caches produced during the regular render and the fact that there can be not real nonce value and other trivialities.